Only need to change if...else...then
Currently if and else push pointers to branch, functions on the stack.
Thats a bad idea. The reason they do this is because it seemed like a
good idea at the time. Their is a small but subtle difference between:
- if...then and
- if...else...then

The first variant requires then to resolve orig pushed by if to branch
if equal whereas the second variant requires an unconditional branch.
Pushing an additional pointer to the needed branch, function to the
stack solved that problem but seems to be a cause troubles when mixing
if...else...then with begin...while constructs.

So I guess we got two other options left?
1.) Make if...else...then only use unconditional branches which would
end up costing additional 4 bytes for every if since we'd need something
like:
	beq pc+4
	orig
	... code following if

2.) Use the LSB of orig to encode whether we'd like to branch
conditionally or unconditionally. Lets try this option first.